splitting memlog to .h and .c
